fix(cli): resolve workspace dependencies in sync-pieces command
The sync-pieces CLI was copying package.json to dist as-is before running
npm pack, leaving workspace:* references unresolved in the uploaded archive.
Extract shared workspace resolution utilities and apply them in buildPiece.
