+++ tools/scripts/utils/publish-npm-package.ts
@@ -1,66 +1,10 @@
 import assert from 'node:assert'
 import { argv } from 'node:process'
 import { execSync } from 'node:child_process'
-import { existsSync, readFileSync, writeFileSync, readdirSync } from 'node:fs'
-import { join } from 'node:path'
+import { existsSync, readFileSync, writeFileSync } from 'node:fs'
 import { readPackageJson } from './files'
 import { packagePrePublishChecks } from './package-pre-publish-checks'
-
-function buildWorkspaceVersionMap(): Map<string, string> {
-  const versionMap = new Map<string, string>()
-  const rootPkg = JSON.parse(readFileSync('package.json').toString())
-  const workspacePatterns: string[] = rootPkg.workspaces ?? []
-
-  for (const pattern of workspacePatterns) {
-    if (pattern.endsWith('/*')) {
-      const dir = pattern.slice(0, -2)
-      if (!existsSync(dir)) {
-        continue
-      }
-      for (const entry of readdirSync(dir, { withFileTypes: true })) {
-        if (entry.isDirectory()) {
-          const pkgPath = join(dir, entry.name, 'package.json')
-          if (existsSync(pkgPath)) {
-            const pkg = JSON.parse(readFileSync(pkgPath).toString())
-            versionMap.set(pkg.name, pkg.version)
-          }
-        }
-      }
-    } else {
-      const pkgPath = join(pattern, 'package.json')
-      if (existsSync(pkgPath)) {
-        const pkg = JSON.parse(readFileSync(pkgPath).toString())
-        versionMap.set(pkg.name, pkg.version)
-      }
-    }
-  }
-
-  return versionMap
-}
-
-function resolveWorkspaceDependencies(
-  deps: Record<string, string> | undefined,
-  versionMap: Map<string, string>,
-): Record<string, string> | undefined {
-  if (!deps) {
-    return deps
-  }
-  const resolved: Record<string, string> = {}
-  for (const [name, version] of Object.entries(deps)) {
-    if (version.startsWith('workspace:')) {
-      const resolvedVersion = versionMap.get(name)
-      if (resolvedVersion) {
-        resolved[name] = resolvedVersion
-        console.info(`[publishPackage] resolved ${name}: ${version} -> ${resolvedVersion}`)
-      } else {
-        throw new Error(`[publishPackage] failed to resolve workspace dependency ${name}: ${version}. Package not found in workspace.`)
-      }
-    } else {
-      resolved[name] = version
-    }
-  }
-  return resolved
-}
+import { buildWorkspaceVersionMap, resolveWorkspaceDependencies } from './workspace-utils'
 
 function assertNoUnresolvedWorkspaceDeps(packageJsonPath: string): void {
   const json = JSON.parse(readFileSync(packageJsonPath).toString())
@@ -104,7 +48,7 @@ export const publishNpmPackage = async (path: string): Promise<void> => {
   const { version } = await readPackageJson(path)
 
   // Update version and resolve workspace dependencies in dist package.json before publishing
-  const versionMap = buildWorkspaceVersionMap()
+  const versionMap = buildWorkspaceVersionMap(process.cwd())
   const json = JSON.parse(readFileSync(`${outputPath}/package.json`).toString())
   json.version = version
   json.main = './src/index.js'
